+++ examples/arrow-function-component.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const TestComponent = (props) => {
+    return (
+        <>
+            {/* Add your component content here */}
+        </>
+    );
+}
+
+export default TestComponent;