+++ examples/class-component.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+class TestComponent extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {};
+    }
+
+    render() {
+        return (
+            <>
+                {/* Add your component content here */}
+            </>
+        );
+    }
+}
+
+export default TestComponent;