+++ examples/forwardref-component.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { forwardRef } from 'react';
+
+const TestComponent = forwardRef((props, ref) => {
+    return (
+        <div ref={ref}>
+            {/* Add your component content here */}
+        </div>
+    );
+});
+
+export default TestComponent;