+++ examples/memoized-component.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { memo } from 'react';
+
+const TestComponent = memo((props) => {
+    return (
+        <>
+            {/* Add your component content here */}
+        </>
+    );
+});
+
+export default TestComponent;