+++ examples/typescript-functional-component.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+interface Props {}
+
+function TestComponent(props: Props) {
+    return (
+        <>
+            {/* Add your component content here */}
+        </>
+    );
+}
+
+export default TestComponent;